requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-ollama>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
mcp>=1.0.0
fastmcp>=0.1.0
gitpython>=3.1.40
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
